refactor: move classes to appropriate packages
