+++ src/main/java/org/simulation/action/SpawnInitialEntitiesAction.java
@@ -14,6 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
+// TODO: refactor
 public class SpawnInitialEntitiesAction implements Action {
 
     private static final int EXTRA_TREES_COUNT = 10;
