--- src/main/java/org/simulation/console/Menu.java
@@ -1,6 +1,0 @@
-package org.simulation.console;
-
-public interface Menu {
-
-    void start();
-}
