+++ src/main/java/org/simulation/factory/SimulationFactoryImpl.java
@@ -5,8 +5,8 @@ import org.simulation.action.MoveEntityAction;
 import org.simulation.action.RandomFreePositionGenerator;
 import org.simulation.action.RespawnEntitiesAction;
 import org.simulation.action.SpawnInitialEntitiesAction;
-import org.simulation.console.renderer.HintRenderer;
-import org.simulation.console.renderer.MapRenderer;
+import org.simulation.ui.console.renderer.HintRenderer;
+import org.simulation.ui.console.renderer.MapRenderer;
 import org.simulation.game.GameMap;
 import org.simulation.game.Simulation;
 import org.simulation.game.SimulationMapConfig;
