+++ src/main/java/org/simulation/ui/console/input/ConsoleInputProvider.java
@@ -1,4 +1,4 @@
-package org.simulation.console;
+package org.simulation.ui.console.input;
 
 import java.util.Objects;
 import java.util.Scanner;
