+++ src/main/java/org/simulation/ui/console/menu/ConsoleControlMenu.java
@@ -1,7 +1,8 @@
-package org.simulation.console;
+package org.simulation.ui.console.menu;
 
-import org.simulation.console.renderer.ControlMenuRenderer;
+import org.simulation.ui.console.renderer.ControlMenuRenderer;
 import org.simulation.game.runner.Runner;
+import org.simulation.ui.console.input.InputProvider;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
