+++ src/main/java/org/simulation/ui/console/menu/Menu.java
@@ -1,0 +1,7 @@
+package org.simulation.ui.console.menu;
+
+// TODO: refactor all menu & renderer
+public interface Menu {
+
+    void start();
+}
