+++ src/main/java/org/simulation/ui/console/renderer/MapRenderer.java
@@ -1,4 +1,4 @@
-package org.simulation.console.renderer;
+package org.simulation.ui.console.renderer;
 
 import org.simulation.game.GameMap;
